Added SDK support for plugins that can index their own audio content.
- Added IIndexerSource interface: plugins will be able to implement this
  interface to add tracks to the library that will be indexed and
  maintained like all other tracks
- Added IIndexerWriter interface: IIndexerSource plugins will use this
  interface to add/remove/update track info with the main app
- Added IIndexerNotifier: interface used to notify the app that it needs
  to be re-indexed.
- Added "source_id" , "external_id", and "visible" column to the tracks
  table, with appropriate indexes.
- Updated all queries to take "source_id" and "visible" into account
- Extracted TrackMetadataQuery from TrackList. Never should have been
  there to begin with, but was due to some technical limitations that no
  longer exist.
- Fixed a really old indexer bug where the reported number of file scanned
  was not accurate. Strange this wasn't noticed before.
- Added "SetIndexerNotifier" injection method
- Fixed a bug in TrackMetadataQuery -- it was unnecessarily relying upon
  the paths table, which was causing query errors when the table was
  empty.
- Use a cache for local file paths instead of requiring an async round
  trip when adding/removing
- Remove use of manual "ANALYZE", it was causing strange performance
  degradation issues. Instead, move to a set of PRAGMAs that instructs
  sqlite to run these optimizations when necessary.
- Updated ::GetInt32 and ::GetUint32 return types to be explicit as to play
  more nicely with clang -- may as well, we're updating the SDK version
  anyway.

+++ src/core/sdk/ITrack.h
@@ -42,8 +42,8 @@ namespace musik { namespace core { namespace sdk {
             virtual int GetValue(const char* key, char* dst, int size) = 0;
             virtual unsigned long long GetUint64(const char* key, unsigned long long defaultValue = 0ULL) = 0;
             virtual long long GetInt64(const char* key, long long defaultValue = 0LL) = 0;
-            virtual unsigned long GetUint32(const char* key, unsigned long defaultValue = 0) = 0;
-            virtual long GetInt32(const char* key, unsigned int defaultValue = 0) = 0;
+            virtual unsigned int GetUint32(const char* key, unsigned long defaultValue = 0) = 0;
+            virtual int GetInt32(const char* key, unsigned int defaultValue = 0) = 0;
             virtual double GetDouble(const char* key, double defaultValue = 0.0f) = 0;
             virtual int Uri(char* dst, int size) = 0;
     };
